Allow for sorted view of the data, based on State and Columns
The current `Table` accepts the list of data in the view function and then
renders the table to the UI after performing a sort of the data.

The current behaviour is completely valid, and the use-case defined
within Table.elm is sound. We have a requirement to use the data, but
have the data be in the same sorted ordering as the data within the `Table`.

Instead of exposing several of the internal structures of the `Table`, it
seems cleaner to allow for a `view` function that would return a new
"view" of the provided data: one where the data is sorted based on the
provided `Table.State` and `Column` definitions. It should be stressed that the
function _does nothing other than sort the data_. Better schemes to achieve
this will be used as soon as they present themselves, and the current
implementation is strictly nothing more than exposure of the internal
sort function.